refactor(engine): Redesign condition evaluation to use Strategy pattern
The previous if/elif chain in the inference engine was becoming
unwieldy and difficult to extend.

This commit refactors the core evaluation logic to use the Strategy
pattern, following the 'Composition over Inheritance' principle. A new
`evaluators` module contains strategy classes for each condition type
(Simple, Conjunctive, Existential, Universal).

This makes the design more modular, extensible, and robust. The
`_run_inference_chain` loop is now a simple polymorphic call, and
adding new condition types in the future will be significantly easier.

This superior design was adopted following an in-depth discussion
about OO principles and the Liskov Substitution Principle.
