+++ src/logic_fabricator/core_types/condition.py
@@ -2,6 +2,12 @@ import json
 import structlog
 
 from .statement import Statement
+from .evaluators import (
+    SimpleConditionEvaluator,
+    ConjunctiveConditionEvaluator,
+    ExistentialConditionEvaluator,
+    UniversalConditionEvaluator,
+)
 
 logger = structlog.get_logger(__name__)
 
@@ -33,6 +39,22 @@ class Condition:
                 "Condition must be one of simple (verb/terms), conjunctive (and_conditions), "
                 "existential (exists_condition), or universal (forall_condition)."
             )
+        
+        self.evaluator = None
+        if is_simple:
+            self.evaluator = SimpleConditionEvaluator()
+        elif is_conjunctive:
+            self.evaluator = ConjunctiveConditionEvaluator()
+        elif is_existential:
+            self.evaluator = ExistentialConditionEvaluator()
+        elif is_universal:
+            self.evaluator = UniversalConditionEvaluator()
+
+    def evaluate(self, known_facts: set["Statement"]) -> dict | None:
+        if self.evaluator:
+            return self.evaluator.evaluate(self, known_facts)
+        return None
+
 
     def __str__(self):
         if self.and_conditions:
@@ -125,182 +147,6 @@ class Condition:
                 verb_synonyms=data.get("verb_synonyms", []),
             )
 
-    def _match_single_condition(self, statement: "Statement") -> dict | None:
-        logger.debug(
-            "Attempting to match single condition", condition=self, statement=statement
-        )
-        verb_matches = (
-            self.verb == statement.verb or statement.verb in self.verb_synonyms
-        )
-        if not verb_matches:
-            logger.debug(
-                "Verb mismatch",
-                condition_verb=self.verb,
-                statement_verb=statement.verb,
-                synonyms=self.verb_synonyms,
-            )
-            return None
+    
 
-        bindings = {}
-        num_cond_terms = len(self.terms)
-        num_stmt_terms = len(statement.terms)
-
-        for i in range(num_cond_terms):
-            cond_term = self.terms[i]
-
-            # Handle wildcard on the last term
-            if cond_term.startswith("*") and i == num_cond_terms - 1:
-                if num_stmt_terms < i:  # Not enough terms to even reach the wildcard
-                    logger.debug(
-                        "Wildcard match failed: not enough statement terms",
-                        condition=self,
-                        statement=statement,
-                    )
-                    return None
-                binding_key = "?" + cond_term[1:]
-                bindings[binding_key] = statement.terms[i:]
-                logger.debug(
-                    "Wildcard matched",
-                    condition=self,
-                    statement=statement,
-                    bindings=bindings,
-                )
-                return bindings  # Wildcard is always last, so we are done.
-
-            # This part runs for non-wildcard terms
-            if i >= num_stmt_terms:  # Not enough statement terms to match the condition
-                logger.debug(
-                    "Term mismatch: not enough statement terms",
-                    condition=self,
-                    statement=statement,
-                )
-                return None
-
-            stmt_term = statement.terms[i]
-            if cond_term.startswith("?"):
-                bindings[cond_term] = stmt_term
-                logger.debug("Variable bound", var=cond_term, value=stmt_term)
-            elif cond_term != stmt_term:
-                logger.debug(
-                    "Literal term mismatch", expected=cond_term, actual=stmt_term
-                )
-                return None
-
-        # If the loop completes, all condition terms matched. If the statement is longer,
-        # that's permissible as per test_rule_applies_with_fewer_condition_terms.
-        if (
-            num_stmt_terms < num_cond_terms
-        ):  # This check seems redundant given the loop logic, but keeping for consistency
-            logger.debug(
-                "Statement shorter than condition terms after loop",
-                condition=self,
-                statement=statement,
-            )
-            return None
-
-        logger.debug(
-            "Single condition matched",
-            condition=self,
-            statement=statement,
-            bindings=bindings,
-        )
-        return bindings
-
-    def _find_consistent_bindings(
-        self,
-        sub_conditions_to_match: list["Condition"],
-        available_statements: list["Statement"],
-        current_bindings: dict,
-    ) -> dict | None:
-        logger.debug(
-            "Attempting to find consistent bindings",
-            sub_conditions_count=len(sub_conditions_to_match),
-            available_statements_count=len(available_statements),
-            current_bindings=current_bindings,
-        )
-        if not sub_conditions_to_match:
-            logger.debug(
-                "All sub-conditions matched, returning bindings",
-                final_bindings=current_bindings,
-            )
-            return (
-                current_bindings  # All sub-conditions matched, return combined bindings
-            )
-
-        sub_condition = sub_conditions_to_match[0]
-        remaining_sub_conditions = sub_conditions_to_match[1:]
-
-        for i, stmt in enumerate(available_statements):
-            logger.debug(
-                "Trying statement for sub-condition",
-                sub_condition=sub_condition,
-                statement=stmt,
-            )
-            sub_bindings = sub_condition._match_single_condition(stmt)
-            if sub_bindings is not None:
-                new_bindings = current_bindings.copy()
-                conflict = False
-                for key, value in sub_bindings.items():
-                    if key in new_bindings and new_bindings[key] != value:
-                        logger.debug(
-                            "Binding conflict detected",
-                            key=key,
-                            existing_value=new_bindings[key],
-                            new_value=value,
-                        )
-                        conflict = True
-                        break
-                    new_bindings[key] = value
-
-                if not conflict:
-                    logger.debug(
-                        "No binding conflict, recursing", new_bindings=new_bindings
-                    )
-                    next_available_statements = (
-                        available_statements[:i] + available_statements[i + 1 :]
-                    )
-                    result = self._find_consistent_bindings(
-                        remaining_sub_conditions,
-                        next_available_statements,
-                        new_bindings,
-                    )
-                    if result is not None:
-                        logger.debug(
-                            "Consistent bindings found in recursion", result=result
-                        )
-                        return result
-        logger.debug("No consistent bindings found for current path")
-        return None
-
-    def matches(self, statements: list["Statement"]) -> dict | None:
-        logger.debug(
-            "Attempting to match condition",
-            condition=self,
-            statements_count=len(statements),
-        )
-        if self.and_conditions is not None:
-            logger.debug(
-                "Matching conjunctive condition", and_conditions=self.and_conditions
-            )
-            result = self._find_consistent_bindings(self.and_conditions, statements, {})
-            if result is not None:
-                logger.debug(
-                    "Conjunctive condition matched", condition=self, bindings=result
-                )
-            else:
-                logger.debug("Conjunctive condition did not match", condition=self)
-            return result
-        else:
-            logger.debug("Matching simple condition", condition=self)
-            for statement in statements:
-                bindings = self._match_single_condition(statement)
-                if bindings is not None:
-                    logger.debug(
-                        "Simple condition matched",
-                        condition=self,
-                        statement=statement,
-                        bindings=bindings,
-                    )
-                    return bindings
-            logger.debug("Simple condition did not match any statement", condition=self)
-            return None  # No match found for any statement
+
